third_party/blink/web_tests/http/tests/websocket/workers/no-onmessage-in-sync-op-expected.txt

WebSocket's message event should not be invoked while we are waiting for the result of a synchronous operation (such as WebSocket.send()) on workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: closeEvent.wasClean is true
PASS PASS: event.length is 9
PASS PASS: events[0] is "Got bufferedAmount: 0"
PASS PASS: events[1] is "Got bufferedAmount: 0"
PASS PASS: events[2] is "Got bufferedAmount: 0"
PASS PASS: events[3] is "Sent message: 1"
PASS PASS: events[4] is "Sent message: 2"
PASS PASS: events[5] is "Sent message: 3"
PASS PASS: events[6] is "Received message: 1"
PASS PASS: events[7] is "Received message: 2"
PASS PASS: events[8] is "Received message: 3"
DONE
PASS successfullyParsed is true

TEST COMPLETE
